+++ src/SecucardConnect/Product/General/ContractsService.php
@@ -6,6 +6,7 @@ use GuzzleHttp\Exception\GuzzleException;
 use SecucardConnect\Client\ApiError;
 use SecucardConnect\Client\AuthError;
 use SecucardConnect\Client\ClientError;
+use SecucardConnect\Client\MissingParamsError;
 use SecucardConnect\Client\ProductService;
 use SecucardConnect\Product\General\Model\PaymentMethodsRequestParams;
 use SecucardConnect\Product\Smart\Model\PaymentWizardContractOptions;
@@ -45,5 +46,27 @@ class ContractsService extends ProductService
     public function getPaymentWizardOptions($contract_id)
     {
         return $this->getWithAction($contract_id, 'iframeOptions', null, null, PaymentWizardContractOptions::class);
+    }
+
+    /**
+     * Remove the accrual flag from all payment transactions of the given contract
+     * (this will be done in the background,
+     *  so the response TRUE will only indicate the the API has accepted this request)
+     *
+     * @param string $contractId The general contract id
+     * @return bool
+     * @throws ApiError
+     * @throws AuthError
+     * @throws ClientError
+     * @throws GuzzleException
+     * @throws MissingParamsError
+     */
+    public function revokeAccrual($contractId)
+    {
+        if (empty($contractId)) {
+            throw new MissingParamsError('contractId', __METHOD__);
+        }
+
+        return (bool)$this->execute($contractId, 'revokeAccrual');
     }
 }