+++ src/SecucardConnect/Product/Payment/ContainersService.php
@@ -1,13 +1,9 @@
 <?php
+/** @noinspection PhpUnused */
 
 namespace SecucardConnect\Product\Payment;
 
-use GuzzleHttp\Exception\GuzzleException;
-use SecucardConnect\Client\ApiError;
-use SecucardConnect\Client\AuthError;
-use SecucardConnect\Client\ClientError;
 use SecucardConnect\Client\ProductService;
-use SecucardConnect\Product\Common\Model\BaseCollection;
 
 /**
  * Operations for the payment/containers resource.