+++ src/SecucardConnect/Product/Payment/ContractsService.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnused */
 
 namespace SecucardConnect\Product\Payment;
 
@@ -6,6 +7,7 @@ use GuzzleHttp\Exception\GuzzleException;
 use SecucardConnect\Client\ApiError;
 use SecucardConnect\Client\AuthError;
 use SecucardConnect\Client\ClientError;
+use SecucardConnect\Client\MissingParamsError;
 use SecucardConnect\Client\ProductService;
 use SecucardConnect\Product\Payment\Model\CloneParams;
 use SecucardConnect\Product\Payment\Model\CreateSubContractRequest;
@@ -24,7 +26,7 @@ class ContractsService extends ProductService
      *
      * @param string $contractId The id of the parent contract.
      * @param CloneParams $param The parameters for cloning.
-     * @return Contract
+     * @return \SecucardConnect\Product\Payment\Model\Contract
      * @throws GuzzleException
      * @throws ApiError
      * @throws AuthError
@@ -54,7 +56,7 @@ class ContractsService extends ProductService
      * Clones the contract of the current user according to the given parameters and returns the contract.
      *
      * @param CloneParams $param The parameters for cloning.
-     * @return Contract
+     * @return \SecucardConnect\Product\Payment\Model\Contract
      * @throws GuzzleException
      * @throws ApiError
      * @throws AuthError
@@ -83,5 +85,27 @@ class ContractsService extends ProductService
         }
 
         return $this->execute($contract_id, 'requestId', null, $param, CreateSubContractResponse::class);
+    }
+
+    /**
+     * Remove the accrual flag from all payment transactions of the given contract
+     * (this will be done in the background,
+     *  so the response TRUE will only indicate the the API has accepted this request)
+     *
+     * @param string $contractId The payment contract id
+     * @return bool
+     * @throws ApiError
+     * @throws AuthError
+     * @throws ClientError
+     * @throws GuzzleException
+     * @throws MissingParamsError
+     */
+    public function revokeAccrual($contractId)
+    {
+        if (empty($contractId)) {
+            throw new MissingParamsError('contractId', __METHOD__);
+        }
+
+        return (bool)$this->execute($contractId, 'revokeAccrual');
     }
 }