+++ src/SecucardConnect/Product/Payment/Model/CrowdFundingDataOpenDetail.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnused */
 
 namespace SecucardConnect\Product\Payment\Model;
 
@@ -12,4 +13,29 @@ class CrowdFundingDataOpenDetail
      * @var int
      */
     public $total = 0;
+
+    /**
+     * @var int|null
+     */
+    public $debit = null;
+
+    /**
+     * @var int|null
+     */
+    public $credit_card = null;
+
+    /**
+     * @var int|null
+     */
+    public $prepay = null;
+
+    /**
+     * @var int|null
+     */
+    public $sofort = null;
+
+    /**
+     * @var int|null
+     */
+    public $twint = null;
 }