+++ src/SecucardConnect/Product/Services/UploadidentsService.php
@@ -2,12 +2,7 @@
 
 namespace SecucardConnect\Product\Services;
 
-use GuzzleHttp\Exception\GuzzleException;
-use SecucardConnect\Client\ApiError;
-use SecucardConnect\Client\AuthError;
-use SecucardConnect\Client\ClientError;
 use SecucardConnect\Client\ProductService;
-use SecucardConnect\Product\Services\Model\IdentCase;
 
 /**
  * Operations for the services/uploadidents resource.